fix: resolving the conflits
