+++ src/app/register/successfully/page.tsx
@@ -1,7 +1,5 @@
-"use client";
-import SectionTitle from "@/components/SectionTitle/SectionTitle";
+import SuccessfullyForm from "@/components/SuccessfullyForm";
 import Image from "next/image";
-import { useRouter } from "next/navigation";
 
 export const metadata = {
    title: "Bora Rachar | Tudo certo com seu cadastro",
@@ -9,7 +7,6 @@ export const metadata = {
 };
 
 export default function Sucessfully() {
-   const router = useRouter();
    return (
       <section className="flex flex-col lg:flex-row-reverse lg:items-center lg:max-h-screen lg:overflow-y-hidden">
          <figure className="w-full h-[475px] bg-[#FFFFFF] hidden overflow-hidden lg:overflow-visible md:flex md:items-center md:justify-center relative lg:h-[100vh] lg:flex-1">
@@ -22,35 +19,7 @@ export default function Sucessfully() {
             />
             <div className="absolute bottom-0 w-full h-1 bg-secondary lg:top-0 lg:left-0 lg:h-full lg:w-1"></div>
          </figure>
-         <section className="mx-4 sm:w-[80%] sm:mx-auto sm:pb-10 md:flex md:items-center md:justify-center md:min-w-[full] md:pb-0 md:flex-1 lg:items-start lg:min-h-[90vh]">
-            <div className="md:flex md:flex-col md:items-start lg:items-start">
-               <figure className="flex items-center justify-center w-full mt-8 lg:mt-0 lg:justify-start md:items-start md:justify-start">
-                  <Image
-                     src="/assets/bora-rachar-logo.svg"
-                     alt="Imagem da logomarca do Bora Rachar"
-                     width={280}
-                     height={100}
-                     className="mx-auto md:mx-0 md:w-[100px] md:h-[40px]"
-                  />
-               </figure>
-               <SectionTitle
-                  title="Cadastro realizado com sucesso!"
-                  customClassesNames={{
-                     title: "text-[32px] font-black sm:text-[32px] ",
-                     header: "mt-8 sm:mt-4",
-                  }}
-               />
-               <p className="text-md text-[#637381] lg:max-w-[50ch]">
-                  Seu email foi verificado e está tudo certo com a sua conta!
-               </p>
-               <button
-                  onClick={() => router.push("/login")}
-                  className="text-lg font-medium w-full text-white bg-[#724FD8] py-[10px] rounded-md cursor-pointer mt-6"
-               >
-                  Entrar na minha conta
-               </button>
-            </div>
-         </section>
+         <SuccessfullyForm />
       </section>
    );
 }
